thebookstore/requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
certifi==2017.11.5
chardet==3.0.4
decorator==4.2.1
Django==1.11
django-allauth==0.34.0
django-bootstrap3==9.1.0
django-countries==5.0
django-crispy-forms==1.7.0
enum34==1.1.6
idna==2.6
ipdb==0.10.3
ipython==5.5.0
ipython-genutils==0.2.0
oauthlib==2.0.6
pathlib2==2.3.0
pbr==3.1.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
python-openid==2.2.5
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
scandir==1.6
simplegeneric==0.8.1
six==1.11.0
stevedore==1.28.0
traitlets==4.3.2
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.8.2
wcwidth==0.1.7
